requirements.txt
# Core dependencies
python-dotenv==1.0.0
customtkinter==5.2.0
Pillow==10.0.0

# API Integration
openai==1.3.0
requests==2.31.0

# Audio processing
pydub==0.25.1
SpeechRecognition==3.10.0

# Document generation
python-docx==0.8.11
fpdf2==2.7.4

# Data handling
numpy